Reduced work done by the `get_user_posts` function.
The `get_user_posts` function was creating the dictionary for the entry even if it wasn't by the right author, then it was just not adding the dictionary to the list when the author didn't match.  I changed the order here so that it would check the author first, and then only do the work if it was the author it cared about.