Workaround the bad default node values in Steem().
Added "api.steemit.com" to the Steem() constructors.  I assume the
underlying library will get an update that makes this un-necessary.  It
would probably be a good idea to create a standard configuration option
for these and save it there.  At least, that would avoid the headache
that a lot of different projects are having right now.