This should fix the gateway timeouts.
Added a list of Steem nodes to the settings object.

Eliminated the global Steem object in favor of only creating local Steem
objects for the connections to steem.

Made the limit on the number of users posts a default argument to
`get_user_posts` so that it can later be made a preference.

Narrowed the scope of the metadata variable to just the region in which
it is accessed.